tests: better test coverage
