+++ tests/cassettes/TestCarrierAccountUpdate.yaml
@@ -1,0 +1,187 @@
+---
+version: 1
+interactions:
+- request:
+    body: '{"carrier_account":{"type":"DhlEcsAccount","credentials":{"client_id":"123456","client_secret":"123abc","distribution_center":"USLAX1","pickup_id":"123456"},"test_credentials":{"client_id":"123456","client_secret":"123abc","distribution_center":"USLAX1","pickup_id":"123456"}}}'
+    form: {}
+    headers:
+      Authorization:
+      - REDACTED
+      Content-Type:
+      - application/json
+      User-Agent:
+      - REDACTED
+    url: https://api.easypost.com/v2/carrier_accounts
+    method: POST
+  response:
+    body: '{"billing_type":"carrier","clone":false,"created_at":"2022-09-19T21:09:16Z","credentials":{},"description":"DHL
+      eCommerce Solutions Account","fields":{"credentials":{"client_id":{"label":"DHL
+      eCommerce Solutions client ID","value":"123456","visibility":"visible"},"client_secret":{"label":"DHL
+      eCommerce Solutions client secret","value":"********","visibility":"password"},"distribution_center":{"label":"DHL
+      eCommerce Solutions distribution center","value":"USLAX1","visibility":"visible"},"pickup_id":{"label":"DHL
+      eCommerce Solutions pickup ID","value":"123456","visibility":"visible"}},"test_credentials":{"client_id":{"label":"Test
+      DHL eCommerce Solutions client ID","value":"123456","visibility":"visible"},"client_secret":{"label":"Test
+      DHL eCommerce Solutions client secret","value":"********","visibility":"password"},"distribution_center":{"label":"Test
+      DHL eCommerce Solutions distribution center","value":"USLAX1","visibility":"visible"},"pickup_id":{"label":"Test
+      DHL eCommerce Solutions pickup ID","value":"123456","visibility":"visible"}}},"id":"ca_145f3b337d7a44149ff677fa6dfbdf40","logo":null,"object":"CarrierAccount","readable":"DHL
+      eCommerce Solutions","reference":null,"test_credentials":{},"type":"DhlEcsAccount","updated_at":"2022-09-19T21:09:16Z"}'
+    headers:
+      Cache-Control:
+      - private, no-cache, no-store
+      Content-Type:
+      - application/json; charset=utf-8
+      Etag:
+      - W/"e93da959f4b0b70a9e81ea828a5b1779"
+      Expires:
+      - "0"
+      Pragma:
+      - no-cache
+      Referrer-Policy:
+      - strict-origin-when-cross-origin
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      X-Backend:
+      - easypost
+      X-Content-Type-Options:
+      - nosniff
+      X-Download-Options:
+      - noopen
+      X-Ep-Request-Uuid:
+      - dcb58cf26328da7ce7886e22000a0969
+      X-Frame-Options:
+      - SAMEORIGIN
+      X-Node:
+      - bigweb8nuq
+      X-Permitted-Cross-Domain-Policies:
+      - none
+      X-Proxied:
+      - intlb1nuq 29913d444b
+      - extlb1nuq 29913d444b
+      X-Runtime:
+      - "0.123143"
+      X-Version-Label:
+      - easypost-202209191945-2356362378-master
+      X-Xss-Protection:
+      - 1; mode=block
+    status: 201 Created
+    code: 201
+    duration: ""
+- request:
+    body: '{"carrier_account":{"id":"ca_145f3b337d7a44149ff677fa6dfbdf40","description":"My
+      custom description","credentials":null,"test_credentials":null}}'
+    form: {}
+    headers:
+      Authorization:
+      - REDACTED
+      Content-Type:
+      - application/json
+      User-Agent:
+      - REDACTED
+    url: https://api.easypost.com/v2/carrier_accounts/ca_145f3b337d7a44149ff677fa6dfbdf40
+    method: PATCH
+  response:
+    body: '{"billing_type":"carrier","clone":false,"created_at":"2022-09-19T21:09:16Z","credentials":{},"description":"My
+      custom description","fields":{"credentials":{"client_id":{"label":"DHL eCommerce
+      Solutions client ID","value":"123456","visibility":"visible"},"client_secret":{"label":"DHL
+      eCommerce Solutions client secret","value":"********","visibility":"password"},"distribution_center":{"label":"DHL
+      eCommerce Solutions distribution center","value":"USLAX1","visibility":"visible"},"pickup_id":{"label":"DHL
+      eCommerce Solutions pickup ID","value":"123456","visibility":"visible"}},"test_credentials":{"client_id":{"label":"Test
+      DHL eCommerce Solutions client ID","value":"123456","visibility":"visible"},"client_secret":{"label":"Test
+      DHL eCommerce Solutions client secret","value":"********","visibility":"password"},"distribution_center":{"label":"Test
+      DHL eCommerce Solutions distribution center","value":"USLAX1","visibility":"visible"},"pickup_id":{"label":"Test
+      DHL eCommerce Solutions pickup ID","value":"123456","visibility":"visible"}}},"id":"ca_145f3b337d7a44149ff677fa6dfbdf40","logo":null,"object":"CarrierAccount","readable":"DHL
+      eCommerce Solutions","reference":null,"test_credentials":{},"type":"DhlEcsAccount","updated_at":"2022-09-19T21:09:16Z"}'
+    headers:
+      Cache-Control:
+      - private, no-cache, no-store
+      Content-Type:
+      - application/json; charset=utf-8
+      Etag:
+      - W/"e521cc683f35a528781eb123c633be0b"
+      Expires:
+      - "0"
+      Pragma:
+      - no-cache
+      Referrer-Policy:
+      - strict-origin-when-cross-origin
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      X-Backend:
+      - easypost
+      X-Content-Type-Options:
+      - nosniff
+      X-Download-Options:
+      - noopen
+      X-Ep-Request-Uuid:
+      - dcb58cf26328da7ce7886e22000a0977
+      X-Frame-Options:
+      - SAMEORIGIN
+      X-Node:
+      - bigweb5nuq
+      X-Permitted-Cross-Domain-Policies:
+      - none
+      X-Proxied:
+      - intlb1nuq 29913d444b
+      - extlb1nuq 29913d444b
+      X-Runtime:
+      - "0.142906"
+      X-Version-Label:
+      - easypost-202209191945-2356362378-master
+      X-Xss-Protection:
+      - 1; mode=block
+    status: 200 OK
+    code: 200
+    duration: ""
+- request:
+    body: ""
+    form: {}
+    headers:
+      Authorization:
+      - REDACTED
+      User-Agent:
+      - REDACTED
+    url: https://api.easypost.com/v2/carrier_accounts/ca_145f3b337d7a44149ff677fa6dfbdf40
+    method: DELETE
+  response:
+    body: '{}'
+    headers:
+      Cache-Control:
+      - private, no-cache, no-store
+      Content-Type:
+      - application/json; charset=utf-8
+      Etag:
+      - W/"44136fa355b3678a1146ad16f7e8649e"
+      Expires:
+      - "0"
+      Pragma:
+      - no-cache
+      Referrer-Policy:
+      - strict-origin-when-cross-origin
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      X-Backend:
+      - easypost
+      X-Content-Type-Options:
+      - nosniff
+      X-Download-Options:
+      - noopen
+      X-Ep-Request-Uuid:
+      - dcb58cf26328da7ce7886e22000a0986
+      X-Frame-Options:
+      - SAMEORIGIN
+      X-Node:
+      - bigweb5nuq
+      X-Permitted-Cross-Domain-Policies:
+      - none
+      X-Proxied:
+      - intlb1nuq 29913d444b
+      - extlb1nuq 29913d444b
+      X-Runtime:
+      - "0.121902"
+      X-Version-Label:
+      - easypost-202209191945-2356362378-master
+      X-Xss-Protection:
+      - 1; mode=block
+    status: 200 OK
+    code: 200
+    duration: ""
