+++ tests/cassettes/TestEventRetrievePayload.yaml
@@ -1,0 +1,111 @@
+---
+version: 1
+interactions:
+- request:
+    body: page_size=5
+    form: {}
+    headers:
+      Authorization:
+      - REDACTED
+      Content-Type:
+      - application/x-www-form-urlencoded
+      User-Agent:
+      - REDACTED
+    url: https://api.easypost.com/v2/events
+    method: GET
+  response:
+    body: '{"events":[{"description":"tracker.updated","id":"evt_2f0ff912353711edb1414b058f79ece1","user_id":"user_a76997d126dd488f8ad22d52301300b0","status":"pending","created_at":"2022-09-15T20:44:24.740Z","mode":"test","object":"Event"},{"description":"tracker.updated","id":"evt_0b3fd340353711edb6b90118615e3ebd","user_id":"user_a76997d126dd488f8ad22d52301300b0","status":"pending","created_at":"2022-09-15T20:43:24.655Z","mode":"test","object":"Event"},{"description":"tracker.updated","id":"evt_e7737390353611ed8f6b6f8b91f724e6","user_id":"user_a76997d126dd488f8ad22d52301300b0","status":"pending","created_at":"2022-09-15T20:42:24.596Z","mode":"test","object":"Event"},{"description":"tracker.updated","id":"evt_c3a1e8e8353611edbf5117a095b5ccdc","user_id":"user_a76997d126dd488f8ad22d52301300b0","status":"pending","created_at":"2022-09-15T20:41:24.502Z","mode":"test","object":"Event"},{"description":"tracker.created","id":"evt_b1c07234353611ed90604f7d07c14a3a","user_id":"user_a76997d126dd488f8ad22d52301300b0","status":"pending","created_at":"2022-09-15T20:40:54.503Z","mode":"test","object":"Event"}],"has_more":false}'
+    headers:
+      Cache-Control:
+      - private, no-cache, no-store
+      Content-Type:
+      - application/json; charset=utf-8
+      Etag:
+      - W/"ff9685eb2e90c8955a379520f17f5f09"
+      Expires:
+      - "0"
+      Pragma:
+      - no-cache
+      Referrer-Policy:
+      - strict-origin-when-cross-origin
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      X-Backend:
+      - easypost
+      X-Content-Type-Options:
+      - nosniff
+      X-Download-Options:
+      - noopen
+      X-Ep-Request-Uuid:
+      - dcb58cf26328c9bbe2b9726f00042c6a
+      X-Frame-Options:
+      - SAMEORIGIN
+      X-Node:
+      - bigweb9nuq
+      X-Permitted-Cross-Domain-Policies:
+      - none
+      X-Proxied:
+      - intlb2nuq 6aa9972c1e
+      - extlb1nuq 29913d444b
+      X-Runtime:
+      - "0.450236"
+      X-Version-Label:
+      - easypost-202209191945-2356362378-master
+      X-Xss-Protection:
+      - 1; mode=block
+    status: 200 OK
+    code: 200
+    duration: ""
+- request:
+    body: ""
+    form: {}
+    headers:
+      Authorization:
+      - REDACTED
+      User-Agent:
+      - REDACTED
+    url: https://api.easypost.com/v2/events/evt_2f0ff912353711edb1414b058f79ece1/payloads
+    method: GET
+  response:
+    body: '{"payloads":[]}'
+    headers:
+      Cache-Control:
+      - private, no-cache, no-store
+      Content-Type:
+      - application/json; charset=utf-8
+      Etag:
+      - W/"b5e6dd0a7b95afc23c69e9e95aa4ce47"
+      Expires:
+      - "0"
+      Pragma:
+      - no-cache
+      Referrer-Policy:
+      - strict-origin-when-cross-origin
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      X-Backend:
+      - easypost
+      X-Content-Type-Options:
+      - nosniff
+      X-Download-Options:
+      - noopen
+      X-Ep-Request-Uuid:
+      - dcb58cf26328c9bce2b9726f00042c9c
+      X-Frame-Options:
+      - SAMEORIGIN
+      X-Node:
+      - bigweb5nuq
+      X-Permitted-Cross-Domain-Policies:
+      - none
+      X-Proxied:
+      - intlb2nuq 6aa9972c1e
+      - extlb1nuq 29913d444b
+      X-Runtime:
+      - "0.033525"
+      X-Version-Label:
+      - easypost-202209191945-2356362378-master
+      X-Xss-Protection:
+      - 1; mode=block
+    status: 200 OK
+    code: 200
+    duration: ""
